mobile updates for blog and scroll updates to main page contact. Update for nav click to close menu on mobile
