+++ components/blogcomponents/Header/Header.jsx
@@ -16,9 +16,9 @@ const Header = () => {
     <div className='container mx-auto px-10 mb-8'>
       <div className='border-b w-full inline-block border-lime-500 py-8'>
         <div className='md:float-left block'>
-          <Link href='/blog'>
-            <span className='cursor-pointer font-bold text-4xl text-white'>
-              Blog Home
+          <Link href='/'>
+            <span className='cursor-pointer font-bold text-2xl text-white'>
+              MP
             </span>
           </Link>
         </div>
@@ -28,13 +28,6 @@ const Header = () => {
               <Link key={category.slug} href={`/category/${category.slug}`}>
                 <span className='md:float-right mt-2 align-middle text-white ml-4 font-semibold cursor-pointer'>
                   {category.name}
-                </span>
-              </Link>
-              <Link
-                href='/'
-              >
-                <span className='md:float-right mt-2 align-middle text-white ml-4 font-semibold cursor-pointer'>
-                  Home
                 </span>
               </Link>
             </>
