+++ components/navbar/mobileNav.js
@@ -1,6 +1,8 @@
 import React, { useState } from 'react'
 import { MenuToggle } from './menuToggle';
-import MobileNavMenu from './mobileNavMenu';
+import Link from 'next/link'
+import { NavLink, NavLinks, NavMenuContainer } from './navbar.styles'
+// import MobileNavMenu from './mobileNavMenu';
 
 import { GithubIcon, HamburgerIcon, HamburgerMenu, LinkedinIcon, MenuContainer, TopContainer, TwitterIcon } from './navbar.styles';
 
@@ -12,6 +14,17 @@ const menuVariants = {
     transform: 'translateX(103%)'
   }
 };
+
+const variants = {
+  show: {
+    transform: 'translateX(0)',
+    opacity: 1
+  },
+  hide: {
+    transform: 'translateX(5em)',
+    opacity: 0
+  }
+}
 
 const menuTransition = {
   type: 'spring',
@@ -136,7 +149,150 @@ const MobileNav = (props) => {
             </a>
           </TwitterIcon>
         </TopContainer>
-        <MobileNavMenu isOpen={isOpen} />
+        <NavMenuContainer>
+          <NavLink
+            initial={false}
+            animate={isOpen ? 'show' : 'hide'}
+            onClick={toggleMenu}
+            variants={{
+              show: {
+                ...variants.show,
+                transition: { delay: 0.3, duration: 0.2 },
+              },
+              hide: {
+                ...variants.hide,
+                transition: { delay: 0.05, duration: 0.05 },
+              },
+            }}
+          >
+            <NavLinks
+              onClick={toggleMenu}
+              to='topHome'
+              className='navItem'
+              smooth={true}
+              duration={500}
+              exact='true'
+              offset={-80}
+              spy={true}
+            >
+              <p className='font-bold -ml-4 text-xl'>
+                Home
+              </p>
+            </NavLinks>
+          </NavLink>
+          <NavLink
+            initial={false}
+            animate={isOpen ? 'show' : 'hide'}
+            variants={{
+              show: {
+                ...variants.show,
+                transition: { delay: 0.4, duration: 0.2 },
+              },
+              hide: {
+                ...variants.hide,
+                transition: { delay: 0.1, duration: 0.05 },
+              },
+            }}
+          >
+            <NavLinks
+              onClick={toggleMenu}
+              to='about'
+              className='navItem'
+              smooth={true}
+              duration={500}
+              exact='true'
+              offset={-80}
+              spy={true}
+            >
+              <p className='font-bold -ml-4 text-xl'>
+                About
+              </p>
+            </NavLinks>
+          </NavLink>
+          <NavLink
+            initial={false}
+            animate={isOpen ? 'show' : 'hide'}
+            variants={{
+              show: {
+                ...variants.show,
+                transition: { delay: 0.5, duration: 0.2 },
+              },
+              hide: {
+                ...variants.hide,
+                transition: { delay: 0.15, duration: 0.05 },
+              },
+            }}
+          >
+            <NavLinks
+              onClick={toggleMenu}
+              to='projects'
+              className='navItem'
+              smooth={true}
+              duration={500}
+              exact='true'
+              offset={-80}
+              spy={true}
+            >
+              <p className='font-bold -ml-4 text-xl'>
+                Projects
+              </p>
+            </NavLinks>
+          </NavLink>
+          <NavLink
+            initial={false}
+            animate={isOpen ? 'show' : 'hide'}
+            variants={{
+              show: {
+                ...variants.show,
+                transition: { delay: 0.6, duration: 0.2 },
+              },
+              hide: {
+                ...variants.hide,
+                transition: { delay: 0.2, duration: 0.05 },
+              },
+            }}
+          >
+            <NavLinks
+              onClick={toggleMenu}
+              to='contact'
+              className='navItem'
+              smooth={true}
+              duration={500}
+              exact='true'
+              offset={-80}
+              spy={true}
+            >
+              <p className='font-bold -ml-4 text-xl'>
+                Contact
+              </p>
+            </NavLinks>
+          </NavLink>
+          <NavLink
+            initial={false}
+            animate={isOpen ? 'show' : 'hide'}
+            variants={{
+              show: {
+                ...variants.show,
+                transition: { delay: 0.7, duration: 0.2 },
+              },
+              hide: {
+                ...variants.hide,
+                transition: { delay: 0.25, duration: 0.05 },
+              },
+            }}
+          >
+            <Link
+              className='navItem'
+              href='/blog'
+              target="_blank"
+              rel="noopener noreferrer"
+            >
+              <p className='font-bold ml-0 text-xl'>
+                Blog
+              </p>
+            </Link>
+          </NavLink>
+        </NavMenuContainer>
       </MenuContainer>
     </HamburgerMenu>
   )
